VideoLLaMA2/requirements.txt
absl-py==2.2.2
accelerate==0.26.1
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioread==3.0.1
av==14.3.0
bitsandbytes==0.43.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.14
decorator==4.4.2
decord==0.6.0
deepspeed==0.13.1
distlib==0.3.9
distro==1.9.0
docker-pycreds==0.4.0
einops==0.6.1
einops-exts==0.0.4
eval_type_backport==0.2.2
exceptiongroup==1.3.0
faiss-gpu-cu11==1.11.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
flash-attn==2.5.8
fonttools==4.58.0
fsspec==2025.3.2
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.44
gradio==3.50.0
gradio_client==0.6.1
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.23.4
idna==3.10
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib_metadata==8.7.0
importlib_resources==6.5.2
iopath==0.1.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
latex2mathml==3.78.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.43.0
Markdown==3.8
markdown2==2.5.3
MarkupSafe==2.1.5
matplotlib==3.9.4
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
narwhals==1.38.2
networkx==3.2.1
ninja==1.11.1.4
numba==0.60.0
numpy==1.24.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-ml-py==12.575.51
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
openai==1.33.0
opencv-python==4.5.5.64
orjson==3.10.18
packaging==25.0
pandas==2.2.3
parameterized==0.9.0
peft==0.4.0
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
portalocker==3.1.1
proglog==0.1.12
protobuf==6.30.2
psutil==5.9.8
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pynvml==12.0.0
pyparsing==3.2.3
pysubs2==1.8.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorchvideo==0.1.5
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
safetensors==0.5.3
scenedetect==0.6.3
scikit-learn==1.2.2
scipy==1.13.1
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.27.0
setproctitle==1.3.6
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
spaces==0.29.2
starlette==0.46.2
svgwrite==1.4.3
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.6.0
timm==1.0.3
tokenizers==0.19.1
torch==2.2.0+cu118
torchaudio==2.2.0+cu118
torchvision==0.17.0+cu118
tqdm==4.67.1
transformers==4.42.3
triton==2.2.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
virtualenv==20.28.1
wandb==0.19.11
wavedrom==2.0.3.post3
websockets==11.0.3
Werkzeug==3.1.3
yacs==0.1.8
zipp==3.21.0
